requirements.txt
boto3==1.14.25
gitdb==4.0.5
docker==4.2.2
